# docs/openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "get": {
        "operationId": "Customers_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive code/name search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Exact resource status filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDataListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "post": {
        "operationId": "Customers_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{partyId}": {
      "get": {
        "operationId": "Customers_get",
        "parameters": [
          {
            "name": "partyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetailDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{partyId}/approve": {
      "post": {
        "operationId": "Customers_approve",
        "parameters": [
          {
            "name": "partyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{partyId}/products": {
      "post": {
        "operationId": "Customers_addProduct",
        "parameters": [
          {
            "name": "partyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartyProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartyProductDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{partyId}/products/{productId}/prices": {
      "post": {
        "operationId": "Customers_createPrice",
        "parameters": [
          {
            "name": "partyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductPriceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPriceDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or domain rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated role is not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Tenant-scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource state or uniqueness conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMS Sprint 0 API",
    "description": "Frozen Sprint 0 operational API contract. Tenant context is derived from the bearer token.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductPriceDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "example": 40
          },
          "effectiveFrom": {
            "type": "string",
            "example": "2026-08-18T00:00:00+05:30"
          },
          "effectiveTo": {
            "type": "string"
          }
        },
        "required": [
          "price",
          "effectiveFrom"
        ]
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "partyCode": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "ABC Shop"
          },
          "relationshipType": {
            "type": "string",
            "enum": [
              "SUBSCRIPTION_ROUTE",
              "AD_HOC",
              "WALK_IN"
            ]
          },
          "creationMode": {
            "type": "string",
            "enum": [
              "TEMPORARY",
              "PERMANENT"
            ]
          },
          "mobile": {
            "type": "string"
          },
          "alternateMobile": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "relationshipType",
          "creationMode"
        ]
      },
      "CreatePartyProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "example": 12
          },
          "quantityMode": {
            "type": "string",
            "enum": [
              "FIXED_PLANNED",
              "RETURN_MATCHED",
              "AD_HOC"
            ]
          },
          "defaultQty": {
            "type": "number"
          },
          "forecastQty": {
            "type": "number",
            "description": "Planning-only quantity; never an execution actual"
          },
          "exchangePolicy": {
            "type": "string",
            "enum": [
              "STRICT",
              "ALLOW_CONTAINER_DUE",
              "STAFF_OVERRIDE"
            ]
          }
        },
        "required": [
          "productId",
          "quantityMode"
        ]
      },
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "APPROVED_RECONCILIATION_NOT_FOUND",
              "AUTHENTICATED_USER_INACTIVE",
              "AUTHENTICATION_REQUIRED",
              "CASH_COUNT_STAFF_MISMATCH",
              "CASH_HANDOVER_NOT_FOUND",
              "CASH_VARIANCE_REASON_REQUIRED",
              "CONTAINER_DUE_AUTHORIZATION_REQUIRED",
              "CRITICAL_RECONCILIATION_EXCEPTION",
              "CUSTOMER_NOT_ACTIVE",
              "CUSTOMER_NOT_FOUND",
              "CUSTOMER_NOT_PENDING_APPROVAL",
              "CUSTOMER_PRICE_PERIOD_OVERLAP",
              "DAMAGE_QUANTITY_REQUIRED",
              "DAMAGE_RATE_NOT_CONFIGURED",
              "DAMAGE_RATE_PERIOD_OVERLAP",
              "DAMAGE_RESOLUTION_REQUIRED",
              "DUPLICATE_EVENT_PRODUCT",
              "EMPTY_EVENT_PRODUCT",
              "EMPTY_RETURN_NOT_ALLOWED",
              "EXCESS_EMPTY_RESOLUTION_REQUIRED",
              "EXCHANGE_RATIO_NOT_ALLOWED",
              "EXCHANGE_RATIO_REQUIRED",
              "FIXED_PLANNED_DEFAULT_QTY_REQUIRED",
              "FIXED_PLANNED_PLANNED_QTY_REQUIRED",
              "FOLLOW_UP_NOT_ALLOWED",
              "FOLLOW_UP_QUANTITY_MISMATCH",
              "FOLLOW_UP_REASON_REQUIRED",
              "HANDOVER_EXCEEDS_STAFF_CASH",
              "HANDOVER_RECIPIENT_NOT_FOUND",
              "HTTP_ERROR",
              "INSUFFICIENT_VEHICLE_STOCK",
              "INTERNAL_ERROR",
              "INVALID_ACCESS_TOKEN",
              "INVALID_CASH_HANDOVER_STATUS",
              "INVALID_CREDENTIALS",
              "INVALID_DEFAULT_QTY",
              "INVALID_EFFECTIVE_PERIOD",
              "INVALID_EXCHANGE_RATIO",
              "INVALID_FORECAST_QTY",
              "INVALID_PLANNED_QTY",
              "INVALID_POST_RECONCILIATION_ADJUSTMENT",
              "INVALID_RECONCILIATION_LOCATION",
              "INVALID_RECONCILIATION_STATUS",
              "INVALID_REFERENCE",
              "INVALID_ROUTE_CUSTOMER",
              "INVALID_ROUTE_STOP_PRODUCT",
              "INVALID_STOP_SOURCE",
              "INVALID_TRIP_STAFF_PERIOD",
              "INVALID_TRIP_STATUS",
              "INVENTORY_STATE_CONFIGURATION_MISSING",
              "LATE_EVENT_PAYMENT_NOT_SUPPORTED",
              "NEGATIVE_DERIVED_CASH",
              "NEGATIVE_DERIVED_STOCK",
              "NOT_FOUND",
              "OPEN_RECONCILIATION_EXCEPTION",
              "PARTIAL_STOP_PRODUCT_NOT_FOUND",
              "PARTY_PRODUCT_NOT_CONFIGURED",
              "PARTY_PRODUCT_NOT_FOUND",
              "PENDING_POST_RECONCILIATION_ADJUSTMENT",
              "POST_RECONCILIATION_ADJUSTMENT_NOT_FOUND",
              "PRICE_NOT_CONFIGURED",
              "PRODUCT_NOT_ACTIVE",
              "PRODUCT_NOT_FOUND",
              "PRODUCT_PRICE_PERIOD_OVERLAP",
              "QUANTITY_MODE_MISMATCH",
              "RECONCILIATION_ALREADY_EXISTS",
              "RECONCILIATION_COUNTS_REQUIRED",
              "RECONCILIATION_EXCEPTION_NOT_FOUND",
              "RECONCILIATION_EXCEPTION_NOT_OPEN",
              "RECONCILIATION_NOT_APPROVED",
              "RECONCILIATION_NOT_EDITABLE",
              "RECONCILIATION_NOT_FOUND",
              "RESOURCE_CONFLICT",
              "ROLE_NOT_AUTHORIZED",
              "ROLE_NOT_FOUND",
              "ROUTE_NOT_FOUND",
              "ROUTE_STOP_NOT_FOUND",
              "ROUTE_STOP_ORDER_CONFLICT",
              "ROUTE_STOP_REORDER_NOT_ALLOWED",
              "SELF_APPROVAL_NOT_ALLOWED",
              "SELF_VARIANCE_APPROVAL_NOT_ALLOWED",
              "STAFF_LOGIN_REQUIRED",
              "STAFF_NOT_ACTIVE",
              "STAFF_NOT_ASSIGNED_TO_TRIP",
              "STAFF_NOT_FOUND",
              "STOCK_DIMENSION_NOT_FOUND",
              "STOCK_VARIANCE_REASON_REQUIRED",
              "STOP_ALREADY_HANDLED",
              "STOP_NOT_FOUND",
              "STOP_REORDER_NOT_ALLOWED",
              "TRIP_NOT_COMPLETED",
              "TRIP_NOT_FOUND",
              "TRIP_STAFF_NOT_FOUND",
              "TRIP_STOP_NOT_FOUND",
              "TRIP_STOP_ORDER_CONFLICT",
              "TRIP_STOP_PRODUCT_NOT_FOUND",
              "TRIP_VEHICLE_REQUIRED",
              "UNRESOLVED_TRIP_STOPS",
              "USER_IDENTITY_REQUIRED",
              "USER_NOT_FOUND",
              "VALIDATION_FAILED",
              "VEHICLE_INVENTORY_LOCATION_REQUIRED",
              "VEHICLE_NOT_AVAILABLE",
              "VEHICLE_NOT_FOUND"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CustomerData": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "example": "1"
          },
          "partyCode": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string"
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "customerStatus": {
            "type": "string"
          },
          "createdSource": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "partyId",
          "partyCode",
          "name",
          "relationshipType",
          "mobile",
          "customerStatus",
          "createdSource",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": true
      },
      "PartyProductData": {
        "type": "object",
        "properties": {
          "partyProductId": {
            "type": "string",
            "example": "1"
          },
          "productId": {
            "type": "string",
            "example": "1"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantityMode": {
            "type": "string"
          },
          "defaultQty": {
            "type": "string",
            "nullable": true
          },
          "forecastQty": {
            "type": "string",
            "nullable": true
          },
          "exchangePolicy": {
            "type": "string",
            "nullable": true
          },
          "activeStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "partyProductId",
          "productId",
          "productCode",
          "productName",
          "quantityMode",
          "defaultQty",
          "forecastQty",
          "exchangePolicy",
          "activeStatus",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": true
      },
...
      "CustomerDetailData": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "example": "1"
          },
          "partyCode": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string"
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "customerStatus": {
            "type": "string"
          },
          "createdSource": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "alternateMobile": {
            "type": "string",
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "locality": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "string",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "approvedByUserId": {
            "type": "string",
            "nullable": true
          },
          "approvedAt": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartyProductData"
            }
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerPriceData"
            }
          }
        },
        "required": [
          "partyId",
          "partyCode",
          "name",
          "relationshipType",
          "mobile",
          "customerStatus",
          "createdSource",
          "createdAt",
          "updatedAt",
          "alternateMobile",
          "addressLine1",
          "addressLine2",
          "locality",
          "city",
          "postalCode",
          "latitude",
          "longitude",
          "createdByUserId",
          "approvedByUserId",
          "approvedAt",
          "products",
          "prices"
        ],
        "additionalProperties": true
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "additionalProperties": true
      },
...
      "CustomerDataListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerData"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": true
      },
      "CustomerDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CustomerData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "CustomerDetailDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CustomerDetailData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "PartyProductDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PartyProductData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
      "CustomerPriceDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CustomerPriceData"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": true
      },
...
    }
  },
...
}